+++ gulp-tasks/compress/images.js
@@ -4,9 +4,10 @@
 const imageMin = require('gulp-imagemin');
 const cache = require('gulp-cache');
 const plumber = require('gulp-plumber');
+const size = require('gulp-size');
 
 const options = {
-  optimizationLevel: 3,
+  optimizationLevel: 5,
   progressive: true,
   interlaced: true,
   verbose: true
@@ -14,9 +15,10 @@ const options = {
 
 module.exports.manualRegister = function (gulp, config) {
   gulp.task('compressImages', function() {
-    return gulp.src(`${config.paths.build}/${config.globs.all}`)
+    return gulp.src(`${config.paths.media}/${config.globs.all}`)
       .pipe(plumber())
       .pipe(cache(imageMin(options)))
-      .pipe(gulp.dest(config.paths.build));
+      .pipe(gulp.dest(config.paths.buildMedia))
+      .pipe(size({title: 'images'}));
   });
 };